Fade to Black when caught update
When walker chaser catches the player, the screen will fade to black. Still having issues with the duration of the fade to black and need to duplicate the walker chaser and remove the runner and slow current in files as they've not been updated yet. Primarily wanted to back up the fade to black currently available.