Update for moving level to level
Includes a small box which can be used to move from level to level when pushed by player. I've created a guide video in case anyone needs it, the only way to watch this video is by using the link on Github or Discord, the video is not public accessible online.

Link to guide:
https://youtu.be/euoCuv9seI0